# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chamber Commissions API — سامانه کمیسیون‌های تخصصی
  version: 1.0.0
  description: 'REST API of the commissions platform. All business endpoints require `Authorization: Bearer <accessToken>`.


    Real-time: Socket.IO at `/socket.io` (auth: `{ token }`). Emit `meeting:join` with a meeting id, then receive `meeting.updated`, `attendance.updated`
    (officers), `quorum.updated`, `agenda.updated`, `comment.created`, `vote.opened`, `vote.progress`, `vote.closed`, and personal `notification`
    events.


    Errors: `{ error: { code, message, details? } }` with Persian messages.'
servers:
- url: http://localhost:4000
security:
- bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: &id001
              type: string
            message: *id001
            details:
              type: array
              items:
                type: object
                properties:
                  path: *id001
                  message: *id001
...
paths:
...
  /api/roles:
    get:
      tags:
      - Roles
      summary: نقش‌ها و دسترسی‌های اتاق (?chamberId) — مدیر اتاق
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - Roles
      summary: تعریف نقش جدید
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chamberId: *id004
                title: *id001
                description: *id001
                hasVote: *id003
                capabilities:
                  type: array
                  items: *id001
  /api/roles/options:
    get:
      tags:
      - Roles
      summary: نقش‌های تعریف‌شده اتاق برای انتخاب سمت
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/roles/{key}:
    put:
      tags:
      - Roles
      summary: تعیین دسترسی‌های یک نقش
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chamberId: *id004
                capabilities:
                  type: array
                  items: *id001
      parameters:
      - name: key
        in: path
        required: true
        schema: *id001
    patch:
      tags:
      - Roles
      summary: ویرایش نقش تعریف‌شده
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chamberId: *id004
                title: *id001
                hasVote: *id003
      parameters:
      - name: key
        in: path
        required: true
        schema: *id001
    delete:
      tags:
      - Roles
      summary: حذف نقش تعریف‌شده (?chamberId)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description: 409 role_in_use
      parameters:
      - name: key
        in: path
        required: true
        schema: *id001
  /api/roles/{key}/reset:
    post:
      tags:
      - Roles
      summary: بازگردانی دسترسی پیش‌فرض نقش پایه
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chamberId: *id004
      parameters:
      - name: key
        in: path
        required: true
        schema: *id001
...